correlation count

+++ cpp/src/dmrpt/algo/drpt_global.cpp
@@ -485,10 +485,10 @@ dmrpt::DRPTGlobal::calculate_tree_leaf_correlation() {
 
     for (int tree = 0; tree < this->ntrees; tree++) {
         correlation_matrix[tree] = vector < vector < vector < int>>>(total_leaf_size);
-        final_mapping[tree] = vector < vector < int >> (total_leaf_size);
+        final_mapping[tree] = vector < vector < float >> (total_leaf_size);
         for (int leaf = 0; leaf < total_leaf_size; leaf++) {
             correlation_matrix[tree][leaf] = vector < vector < int >> (this->ntrees);
-            final_mapping[tree][leaf] = vector<int>(ntrees);
+            final_mapping[tree][leaf] = vector<float>(ntrees);
 
             vector <DataPoint> data_points = this->trees_leaf_first_indices[tree][leaf];
